requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
cleverhans==2.1.0
cryptography==2.1.4
cycler==0.10.0
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.6
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==2.2.3
mnist==0.2.2
nose==1.3.7
numpy==1.14.5
pandas==1.0.3
Pillow==7.1.1
praatio==3.7.0
protobuf==3.11.3
pycodestyle==2.5.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
scipy==1.0.0
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.45.0
Werkzeug==1.0.1
wrapt==1.12.1